Configureations for javaToes routing class
