+++ src/main/resources/public/index.html
@@ -18,7 +18,7 @@
                 <div id="bottom">
                     <div>
                     <h1 class="alert alert-success"> right div </h3>
-
+                        <h1 id="status"></h1>
                       
                     </div>
                     <br/><br/>
